feat(name): add utility to get initials for names
