+++ src/utils/strings/index.ts
@@ -1,1 +1,2 @@
 export * from './string';
+export * from './name';
